+++ next.config.ts
@@ -5,6 +5,7 @@ const nextConfig: NextConfig = {
   experimental: {
     // TypeScript 7은 JS Compiler API가 없어 로컬 tsc로 타입체크
     useTypeScriptCli: true,
+    optimizePackageImports: ['lucide-react', 'framer-motion', 'date-fns'],
   },
 };
