+++ src/shared/ui/index.ts
@@ -1,15 +1,14 @@
+/**
+ * 경량 UI만 배럴로 노출합니다.
+ * Three.js / Novel / 에디터 등 무거운 모듈은 직접 경로로 import 하세요.
+ * (배럴을 통한 side-effect preload가 블로그 라우트에 3D 에셋을 끌어오던 문제 방지)
+ */
 export * from './Meteors';
 export * from './AuthSidebar';
-export * from './Loader';
 export * from './modal';
 export * from './Navbar';
-export * from './NovelViewer';
 export * from './Tag';
 export * from './VisitorLogger';
-export * from './AnimatedText';
-export * from './WatchRobot';
-export * from './HelloBot';
-export * from './MoveBot';
 export * from './FloatingActionButton';
 export { Dropdown } from './Dropdown';
 export { Button } from './Button';
