+++ src/widgets/home/index.ts
@@ -1,1 +1,6 @@
-export * from './ui';
+/**
+ * 홈 위젯 public API.
+ * CosmosCursor / 3D 씬은 side-effect preload가 있으므로 필요 시 직접 경로로 import 하세요.
+ */
+export { PortalDiscoveryOverlay } from './ui/PortalDiscoveryOverlay';
+export { PortalEnterTransition } from './ui/PortalEnterTransition';
