+++ src/widgets/post/ui/index.ts
@@ -1,4 +1,7 @@
-export { Macintosh } from './Macintosh';
+/**
+ * 블로그 글 목록/상세용 경량 export만 유지.
+ * Macintosh / ComputerBackground 등 3D 모듈은 직접 경로로 import.
+ */
 export { BlogContent } from './BlogContent';
 export { BlogMainPage } from './BlogMainPage';
 export { default as BlogHeader } from './BlogHeader';
@@ -10,4 +13,3 @@ export { PostsLayoutShell } from './PostsLayoutShell';
 export { BlogSectionTitle } from './BlogSectionTitle';
 export { SearchBar } from './SearchBar';
 export { CategoryTree } from './CategoryTree';
-export * from './ComputerBackground/index';
